+++ Assets/CommonCoreModules/World/ObjectComponents/IReceiveEntityEvents.cs
@@ -18,10 +18,6 @@ namespace CommonCore.World
 
         void BeforeDestroy(); //needed?
 
-        //no save/load, that's handled by another interface
-
-        //TODO CONCEPTUAL entity-related things like "onkilled"?
-
     }
 
     /// <summary>
@@ -32,13 +28,5 @@ namespace CommonCore.World
         void DamageTaken(ActorHitInfo data);
 
         void Killed();
-    }
-
-    /// <summary>
-    /// Interface representing a component on an object that receives actor entity event calls
-    /// </summary>
-    public interface IReceiveActorEntityEvents : IReceiveDamageableEntityEvents
-    {
-
-    }
+    }    
 }